The 'push' event executes the callback even when the app is in the background.
When the app is running in the background (eg. it is visible in the multitasking bar) and receives a notification, the OS alert box pops up. If the user decides to launch the app, push.registerEvent('push', function (push) {}); is called. The 'push' event that is registered should only execute the callback when a push is received while the application is running.